+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Timezone preferences in settings page.
+
+### Changed
+- Send user timezone in register mutation.
+
 ### Fixed
 - Sentry dsn environment variable name.