+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Browser default outline in components.
 - Remove usage of `style: 'unit'` in `NumberFormat`.
+- Issues in model page for mobile viewports.
 
 ## [0.7.1] - 2020-07-20
 ### Fixed