+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Browser default outline in components.
+- Locale popover position overflowing on mobile.
 - Remove usage of `style: 'unit'` in `NumberFormat`.
 - Issues in model page for mobile viewports.
 - Overflow in deck description in home page.