+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Remove usage of `style: 'unit'` in `NumberFormat`.
 - Issues in model page for mobile viewports.
 - Overflow in deck description in home page.
+- Home layout for mobile viewports.
 
 ## [0.7.1] - 2020-07-20
 ### Fixed