+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Add tooltip to homepage toolbar and update styling.
+
 ## [0.7.3] - 2020-07-23
 ### Fixed
 - Incorrect date in study frequency chart of statistics page.